requirements.txt
streamlit
pandas
matplotlib
pyserial
Pillow
plotly
fpdf
